feat(playground): Add WASM execution mode support
